Get AcceptanceTests working
