--- src/NServiceBus.Transport.Msmq.AcceptanceTests/EndpointCustomizationConfigurationExtensions.cs
@@ -1,69 +1,0 @@
-﻿namespace NServiceBus.AcceptanceTests.EndpointTemplates
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reflection;
-    using AcceptanceTesting.Support;
-    using Hosting.Helpers;
-    using ObjectBuilder;
-
-    public static class EndpointCustomizationConfigurationExtensions
-    {
-        public static IEnumerable<Type> GetTypesScopedByTestClass(this EndpointCustomizationConfiguration endpointConfiguration)
-        {
-            var assemblies = new AssemblyScanner().GetScannableAssemblies();
-
-            var types = assemblies.Assemblies
-                //exclude all test types by default
-                .Where(a =>
-                {
-                    var references = a.GetReferencedAssemblies();
-
-                    return references.All(an => an.Name != "nunit.framework");
-                })
-                .SelectMany(a => a.GetTypes());
-
-            types = types.Union(GetNestedTypeRecursive(endpointConfiguration.BuilderType.DeclaringType, endpointConfiguration.BuilderType));
-
-            types = types.Union(endpointConfiguration.TypesToInclude);
-
-            return types.Where(t => !endpointConfiguration.TypesToExclude.Contains(t)).ToList();
-        }
-
-        static IEnumerable<Type> GetNestedTypeRecursive(Type rootType, Type builderType)
-        {
-            if (rootType == null)
-            {
-                throw new InvalidOperationException("Make sure you nest the endpoint infrastructure inside the TestFixture as nested classes");
-            }
-
-            yield return rootType;
-
-            if (typeof(IEndpointConfigurationFactory).IsAssignableFrom(rootType) && rootType != builderType)
-            {
-                yield break;
-            }
-
-            foreach (var nestedType in rootType.GetNestedTypes(BindingFlags.Instance | BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic).SelectMany(t => GetNestedTypeRecursive(t, builderType)))
-            {
-                yield return nestedType;
-            }
-        }
-
-        public static void RegisterComponentsAndInheritanceHierarchy(this EndpointConfiguration builder, RunDescriptor runDescriptor)
-        {
-            builder.RegisterComponents(r => { RegisterInheritanceHierarchyOfContextOnContainer(runDescriptor, r); });
-        }
-
-        static void RegisterInheritanceHierarchyOfContextOnContainer(RunDescriptor runDescriptor, IConfigureComponents r)
-        {
-            var type = runDescriptor.ScenarioContext.GetType();
-            while (type != typeof(object))
-            {
-                r.RegisterSingleton(type, runDescriptor.ScenarioContext);
-                type = type.BaseType;
-            }
-        }
-    }
-}
