+++ src/NServiceBus.Transport.Msmq.AcceptanceTests/NServiceBus.Transport.Msmq.AcceptanceTests.csproj
@@ -2,13 +2,11 @@
 
   <PropertyGroup>
     <TargetFramework>net46</TargetFramework>
-    <RootNamespace>NServiceBus.Transport.Msmq.AcceptanceTests</RootNamespace>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NServiceBus.AcceptanceTesting\NServiceBus.AcceptanceTesting.csproj" />
-    <ProjectReference Include="..\NServiceBus.Core\NServiceBus.Core.csproj" />
+    <ProjectReference Include="..\NserviceBus.Transport.Msmq\NserviceBus.Transport.Msmq.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -19,7 +17,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
-    <PackageReference Include="NUnit" Version="3.6.1" />
+    <PackageReference Include="NServiceBus.AcceptanceTests.Sources" Version="7.0.0-remove-msmq0001" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.7.0" />
   </ItemGroup>
