--- src/NServiceBus.Transport.Msmq.AcceptanceTests/NServiceBusAcceptanceTest.cs
@@ -1,36 +1,0 @@
-namespace NServiceBus.AcceptanceTests
-{
-    using System.Linq;
-    using System.Threading;
-    using AcceptanceTesting.Customization;
-    using NUnit.Framework;
-
-    /// <summary>
-    /// Base class for all the NSB test that sets up our conventions
-    /// </summary>
-    [TestFixture]
-    // ReSharper disable once PartialTypeWithSinglePart
-    public abstract partial class NServiceBusAcceptanceTest
-    {
-        [SetUp]
-        public void SetUp()
-        {
-            Conventions.EndpointNamingConvention = t =>
-            {
-                var classAndEndpoint = t.FullName.Split('.').Last();
-
-                var testName = classAndEndpoint.Split('+').First();
-
-                testName = testName.Replace("When_", "");
-
-                var endpointBuilder = classAndEndpoint.Split('+').Last();
-
-                testName = Thread.CurrentThread.CurrentCulture.TextInfo.ToTitleCase(testName);
-
-                testName = testName.Replace("_", "");
-
-                return testName + "." + endpointBuilder;
-            };
-        }
-    }
-}
