+++ src/NServiceBus.Transport.Msmq.AcceptanceTests/ServerWithNoErrorQueueConfigured.cs
@@ -1,0 +1,55 @@
+﻿namespace NServiceBus.Transport.Msmq.AcceptanceTests
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
+    using AcceptanceTesting.Support;
+    using NServiceBus.AcceptanceTesting.Customization;
+    using NServiceBus.AcceptanceTests.EndpointTemplates;
+    using NServiceBus.Configuration.AdvancedExtensibility;
+    using NServiceBus.Features;
+
+    class ServerWithNoErrorQueueConfigured : IEndpointSetupTemplate
+    {
+        public ServerWithNoErrorQueueConfigured()
+        {
+            typesToInclude = new List<Type>();
+        }
+
+        public ServerWithNoErrorQueueConfigured(List<Type> typesToInclude)
+        {
+            this.typesToInclude = typesToInclude;
+        }
+
+        public async Task<EndpointConfiguration> GetConfiguration(RunDescriptor runDescriptor, EndpointCustomizationConfiguration endpointConfiguration, Action<EndpointConfiguration> configurationBuilderCustomization)
+        {
+            var types = endpointConfiguration.GetTypesScopedByTestClass();
+
+            typesToInclude.AddRange(types);
+
+            var configuration = new EndpointConfiguration(endpointConfiguration.EndpointName);
+
+            configuration.TypesToIncludeInScan(typesToInclude);
+            configuration.EnableInstallers();
+
+            configuration.DisableFeature<TimeoutManager>();
+
+            var recoverability = configuration.Recoverability();
+            recoverability.Delayed(delayed => delayed.NumberOfRetries(0));
+            recoverability.Immediate(immediate => immediate.NumberOfRetries(0));
+
+            await configuration.DefineTransport(runDescriptor, endpointConfiguration).ConfigureAwait(false);
+
+            configuration.RegisterComponentsAndInheritanceHierarchy(runDescriptor);
+
+            await configuration.DefinePersistence(runDescriptor, endpointConfiguration).ConfigureAwait(false);
+
+            configuration.GetSettings().SetDefault("ScaleOut.UseSingleBrokerQueue", true);
+            configurationBuilderCustomization(configuration);
+
+            return configuration;
+        }
+
+        List<Type> typesToInclude;
+    }
+}
