+++ src/NServiceBus.Transport.Msmq.AcceptanceTests/SubscriptionStorage/When_using_subscription_store.cs
@@ -6,7 +6,6 @@
     using NServiceBus.AcceptanceTests;
     using NServiceBus.AcceptanceTests.EndpointTemplates;
     using NUnit.Framework;
-    using Persistence.Legacy;
 
     public class When_using_subscription_store : NServiceBusAcceptanceTest
     {
