+++ src/NServiceBus.Transport.Msmq.sln
@@ -5,11 +5,11 @@ VisualStudioVersion = 15.0.26430.6
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Transport.Msmq.TransportTests", "NServiceBus.Transport.Msmq.TransportTests\NServiceBus.Transport.Msmq.TransportTests.csproj", "{C7255A45-6D42-48BF-8CB5-5CD0FE996B71}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Transport.Msmq.AcceptanceTests", "NServiceBus.Transport.Msmq.AcceptanceTests\NServiceBus.Transport.Msmq.AcceptanceTests.csproj", "{175CCACE-FC7D-474C-B02D-6A505B24B600}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NserviceBus.Transport.Msmq", "NserviceBus.Transport.Msmq\NserviceBus.Transport.Msmq.csproj", "{D60AC41F-E7A9-46C7-9086-87A3D2AD3325}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus.Transport.Msmq.Tests", "NServiceBus.Transport.Msmq.Tests\NServiceBus.Transport.Msmq.Tests.csproj", "{A366459A-1F80-477D-BBFB-5A8C8132F22C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NServiceBus.Transport.Msmq.AcceptanceTests", "NServiceBus.Transport.Msmq.AcceptanceTests\NServiceBus.Transport.Msmq.AcceptanceTests.csproj", "{70EDA9A1-B359-42A5-B12B-97FFFF5466F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,10 +21,6 @@ Global
 		{C7255A45-6D42-48BF-8CB5-5CD0FE996B71}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C7255A45-6D42-48BF-8CB5-5CD0FE996B71}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C7255A45-6D42-48BF-8CB5-5CD0FE996B71}.Release|Any CPU.Build.0 = Release|Any CPU
-		{175CCACE-FC7D-474C-B02D-6A505B24B600}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{175CCACE-FC7D-474C-B02D-6A505B24B600}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{175CCACE-FC7D-474C-B02D-6A505B24B600}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{175CCACE-FC7D-474C-B02D-6A505B24B600}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D60AC41F-E7A9-46C7-9086-87A3D2AD3325}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D60AC41F-E7A9-46C7-9086-87A3D2AD3325}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D60AC41F-E7A9-46C7-9086-87A3D2AD3325}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -33,6 +29,10 @@ Global
 		{A366459A-1F80-477D-BBFB-5A8C8132F22C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A366459A-1F80-477D-BBFB-5A8C8132F22C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A366459A-1F80-477D-BBFB-5A8C8132F22C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70EDA9A1-B359-42A5-B12B-97FFFF5466F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70EDA9A1-B359-42A5-B12B-97FFFF5466F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70EDA9A1-B359-42A5-B12B-97FFFF5466F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70EDA9A1-B359-42A5-B12B-97FFFF5466F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
